Refactor swipe screen to separate deck state, animations, UI and side effects
