+++ frontend/app/_layout.tsx
@@ -7,7 +7,7 @@ export default function RootLayout() {
       <Stack.Screen name="room" options={{ title: "Room" }} />
       <Stack.Screen name="error" options={{ presentation: "modal", headerShown: false }} />
       <Stack.Screen name="join" options={{ title: "Join Room" }} />
-      <Stack.Screen name="swipe" options={{ title: "Placeholder" }} />
+      <Stack.Screen name="swipe/index" options={{ title: "Placeholder" }} />
     </Stack>
   );
 }
