+++ frontend/src/swipe/constants.ts
@@ -1,0 +1,3 @@
+export const SWIPE_THRESHOLD_RATIO = 0.25;
+export const SWIPE_OUT_DURATION_MS = 180;
+export const EDGE_WIDTH = 64;
